+++ NEWS.md
@@ -1,5 +1,8 @@
 # ggfacto development version
 
+## Added : 
+* New argument `reverse_axes` in `ggmca()` to invert left and right or up and down. 
+
 ## Bug correction
 * `mca_interpret` was not working with `MCA(excl = )`
 * in `ggmca`, profiles of answers were plain wrong (because of the unanticipated results of a change in code)